# web/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API Documentation",
    "version" : "1.0"
  },
  "paths" : {
...
    "/order/create" : {
      "post" : {
        "summary" : "createOrder",
        "description" : "createOrder",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "userId" : {
                    "type" : "string",
                    "description" : "用户ID"
                  },
                  "productId" : {
                    "type" : "string",
                    "description" : "商品ID"
                  },
                  "quantity" : {
                    "type" : "integer",
                    "description" : "购买数量"
                  }
                },
                "description" : ""
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "object",
                      "properties" : {
                        "orderId" : {
                          "type" : "string",
                          "description" : "订单ID"
                        },
                        "payUrl" : {
                          "type" : "string",
                          "description" : "支付URL"
                        },
                        "orderStatus" : {
                          "type" : "string",
                          "description" : "订单状态"
                        },
                        "orderStatusDesc" : {
                          "type" : "string",
                          "description" : "订单状态描述"
                        }
                      },
                      "description" : ""
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/order/query" : {
      "get" : {
        "summary" : "queryOrder",
        "description" : "queryOrder",
        "parameters" : [ {
          "name" : "orderId",
          "in" : "query",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "object",
                      "properties" : {
                        "orderId" : {
                          "type" : "string",
                          "description" : "订单ID"
                        },
                        "productId" : {
                          "type" : "string",
                          "description" : "商品ID"
                        },
                        "productName" : {
                          "type" : "string",
                          "description" : "商品名称"
                        },
                        "totalAmount" : {
                          "type" : "number",
                          "description" : "订单金额"
                        },
                        "quantity" : {
                          "type" : "integer",
                          "description" : "购买数量"
                        },
                        "orderStatus" : {
                          "type" : "string",
                          "description" : "订单状态"
                        },
                        "orderStatusDesc" : {
                          "type" : "string",
                          "description" : "订单状态描述"
                        },
                        "orderTime" : {
                          "type" : "string",
                          "description" : "下单时间"
                        },
                        "payUrl" : {
                          "type" : "string",
                          "description" : "支付URL"
                        },
                        "payTime" : {
                          "type" : "string",
                          "description" : "支付时间"
                        }
                      },
                      "description" : ""
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/order/user-orders" : {
      "get" : {
        "summary" : "queryUserOrders",
        "description" : "queryUserOrders",
        "parameters" : [ {
          "name" : "userId",
          "in" : "query",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "array",
                      "description" : "",
                      "items" : {
                        "type" : "object",
                        "properties" : {
                          "orderId" : {
                            "type" : "string",
                            "description" : "订单ID"
                          },
                          "productId" : {
                            "type" : "string",
                            "description" : "商品ID"
                          },
                          "productName" : {
                            "type" : "string",
                            "description" : "商品名称"
                          },
                          "totalAmount" : {
                            "type" : "number",
                            "description" : "订单金额"
                          },
                          "quantity" : {
                            "type" : "integer",
                            "description" : "购买数量"
                          },
                          "orderStatus" : {
                            "type" : "string",
                            "description" : "订单状态"
                          },
                          "orderStatusDesc" : {
                            "type" : "string",
                            "description" : "订单状态描述"
                          },
                          "orderTime" : {
                            "type" : "string",
                            "description" : "下单时间"
                          },
                          "payUrl" : {
                            "type" : "string",
                            "description" : "支付URL"
                          },
                          "payTime" : {
                            "type" : "string",
                            "description" : "支付时间"
                          }
                        },
                        "description" : ""
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}